autorisasi

+++ src/api/notes/handler.js
@@ -19,10 +19,15 @@ class NotesHandler {
         title = 'untitled', body, tags
       } = request.payload;
 
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+
       const noteId = await this._service.addNote({
         title,
         body,
-        tags
+        tags,
+        owner: credentialId,
       });
 
       const response = h.response({
@@ -60,7 +65,11 @@ class NotesHandler {
     }
   }
   async getNotesHandler() {
-    const notes = await this._service.getNotes();
+    const {
+      id: credentialId
+    } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
+
     return {
       status: 'success',
       data: {
@@ -73,7 +82,14 @@ class NotesHandler {
       const {
         id
       } = request.params;
-      const note =await this._service.getNoteById(id);
+
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
+
+      const note = await this._service.getNoteById(id);
       return {
         status: 'success',
         data: {
@@ -104,9 +120,15 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
+
       const {
         id
       } = request.params;
+
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
 
       await this._service.editNoteById(id, request.payload);
 
@@ -139,6 +161,13 @@ class NotesHandler {
       const {
         id
       } = request.params;
+
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+      
+      await this._service.verifyNoteOwner(id, credentialId);
+
       await this._service.deleteNoteById(id);
       return {
         status: 'success',
